+++ Boa.Constrictor.Selenium/Tasks/SwitchFrame.cs
@@ -1,11 +1,14 @@
 ﻿using Boa.Constrictor.Screenplay;
 using OpenQA.Selenium;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Boa.Constrictor.Selenium
 {
     /// <summary>
     /// Switches the frame.
+    /// Start from DefaultContent by default.
     /// </summary>
     public class SwitchFrame : AbstractWebTask
     {
@@ -15,13 +18,12 @@ namespace Boa.Constrictor.Selenium
         /// Private constructor.
         /// (Use static builder methods to construct.)
         /// </summary>
-        /// <param name="locator">The locator.</param>
-        /// <param name="startFromDefaultContent">If true switch to DefaultContent before switching to the locator.</param>
-        /// <param name="useDefaultContent">If true use DefaultContent instead of the locator.</param>
-        private SwitchFrame(IWebLocator locator, bool useDefaultContent, bool startFromDefaultContent = true)
+        /// <param name="locators">The list of locators.</param>
+        /// <param name="useDefaultContent">If true use DefaultContent instead of the locators.</param>
+        private SwitchFrame(List<IWebLocator> locators, bool useDefaultContent)
         {
-            Locator = locator;
-            StartFromDefaultContent = startFromDefaultContent;
+            Locators = locators;
+            StartFromCurrentLocation = false;
             UseDefaultContent = useDefaultContent;
         }
 
@@ -30,17 +32,17 @@ namespace Boa.Constrictor.Selenium
         #region Properties
 
         /// <summary>
-        /// The target Web element's locator.
+        /// The list of target Web element locators.
         /// </summary>
-        public IWebLocator Locator { get; }
+        public List<IWebLocator> Locators { get; }
 
         /// <summary>
-        /// Switch to DefaultContent first before switching to a frame.
+        /// Start from current location instead of switching to DefaultContent first.
         /// </summary>
-        public bool StartFromDefaultContent { get; }
+        public bool StartFromCurrentLocation { get; set; }
 
         /// <summary>
-        /// The DefaultContent is used.
+        /// Switch to DefaultContent instead of a target frame.
         /// </summary>
         public bool UseDefaultContent { get; }
 
@@ -53,20 +55,31 @@ namespace Boa.Constrictor.Selenium
         /// </summary>
         /// <param name="locator">The locator.</param>
         /// <returns></returns>
-        public static SwitchFrame To(IWebLocator locator) => new SwitchFrame(locator, false);
+        public static SwitchFrame To(IWebLocator locator) =>
+            new SwitchFrame(new List<IWebLocator> { locator }, false);
 
         /// <summary>
         /// Constructs the Task object for DefaultContent.
         /// </summary>
         /// <returns></returns>
-        public static SwitchFrame ToDefaultContent() => new SwitchFrame(null, true);
+        public static SwitchFrame ToDefaultContent() => new SwitchFrame(new List<IWebLocator>(), true);
 
         /// <summary>
-        /// Constructs the Task object for the given locator without starting from DefaultContent.
+        /// Constructs the Task object for the given locator list.
         /// </summary>
-        /// <param name="locator">The locator.</param>
+        /// <param name="locators">The list of locators.</param>
         /// <returns></returns>
-        public static SwitchFrame WithoutUsingDefaultContentTo(IWebLocator locator) => new SwitchFrame(locator, false, false);
+        public static SwitchFrame ToNested(List<IWebLocator> locators) => new SwitchFrame(locators, false);
+
+        /// <summary>
+        /// Sets the Task to start from current location instead of DefaultContent.
+        /// </summary>
+        /// <returns></returns>
+        public SwitchFrame AndStartFromCurrentLocation()
+        {
+            StartFromCurrentLocation = true;
+            return this;
+        }
 
         #endregion
 
@@ -79,13 +92,16 @@ namespace Boa.Constrictor.Selenium
         /// <param name="driver">The WebDriver.</param>
         public override void PerformAs(IActor actor, IWebDriver driver)
         {
-            if (StartFromDefaultContent)
+            if (UseDefaultContent || !StartFromCurrentLocation)
                 driver.SwitchTo().DefaultContent();
 
             if (!UseDefaultContent)
             {
-                actor.WaitsUntil(Existence.Of(Locator), IsEqualTo.True());
-                driver.SwitchTo().Frame(Locator.FindElement(driver));
+                foreach (IWebLocator locator in Locators)
+                {
+                    actor.WaitsUntil(Existence.Of(locator), IsEqualTo.True());
+                    driver.SwitchTo().Frame(locator.FindElement(driver));
+                }
             }
         }
 
@@ -95,8 +111,8 @@ namespace Boa.Constrictor.Selenium
         /// <param name="obj">The other object.</param>
         public override bool Equals(object obj) =>
             obj is SwitchFrame frame &&
-            Locator.Equals(frame.Locator) &&
-            StartFromDefaultContent == frame.StartFromDefaultContent &&
+            Locators.Equals(frame.Locators) &&
+            StartFromCurrentLocation == frame.StartFromCurrentLocation &&
             UseDefaultContent == frame.UseDefaultContent;
 
         /// <summary>
@@ -104,7 +120,7 @@ namespace Boa.Constrictor.Selenium
         /// </summary>
         /// <returns></returns>
         public override int GetHashCode() =>
-            HashCode.Combine(GetType(), Locator, StartFromDefaultContent, UseDefaultContent);
+            HashCode.Combine(GetType(), Locators, StartFromCurrentLocation, UseDefaultContent);
 
         /// <summary>
         /// Returns a description of the Task.
@@ -113,7 +129,7 @@ namespace Boa.Constrictor.Selenium
         public override string ToString() =>
             UseDefaultContent
                 ? "switch frame to DefaultContent"
-                : $"switch frame to '{Locator.Description}'";
+                : $"switch frame to '{Locators.Last().Description}'";
 
         #endregion
     }